+++ Editor/R2APISupport/PropertyDrawers/Deprecated/AddressReferencedAssetDrawer.cs
@@ -42,7 +42,7 @@ namespace RoR2.Editor.PropertyDrawers
 
 
             string assetName = null;
-            if(AddressablesPathDictionary.GetInstance().TryGetPathFromGUID(_addressProperty.stringValue, out var path))
+            if(AddressablesPathDictionary.instance.TryGetPathFromGUID(_addressProperty.stringValue, out var path))
             {
                 assetName = System.IO.Path.GetFileNameWithoutExtension(path);
             }
@@ -160,7 +160,7 @@ namespace RoR2.Editor.PropertyDrawers
 
             if(result == null)
             {
-                Debug.LogWarning($"Could not automatically obtain the asset type for AddressReferencedAsset<> in {fieldInfo.DeclaringType.FullName + "." + fieldInfo.Name}");
+                RoR2EKLog.Warning($"Could not automatically obtain the asset type for AddressReferencedAsset<> in {fieldInfo.DeclaringType.FullName + "." + fieldInfo.Name}");
             }
             return result;
         }
@@ -173,7 +173,7 @@ namespace RoR2.Editor.PropertyDrawers
 
         private void ValidateAssetAndAssign(AddressablesPathDropdown.Item item, SerializedProperty addressProperty)
         {
-            if (AddressablesPathDictionary.GetInstance().TryGetGUIDFromPath(item.assetPath, out var guid))
+            if (AddressablesPathDictionary.instance.TryGetGUIDFromPath(item.assetPath, out var guid))
             {
                 addressProperty.stringValue = guid;
                 addressProperty.serializedObject.ApplyModifiedProperties();